bug[TW23845]: Improve processing of ImportActionsViaSpreadsheetBlam

+++ plugins/org.eclipse.osee.ats.rest/src/org/eclipse/osee/ats/rest/internal/util/AtsChangeSet.java
@@ -321,8 +321,10 @@ public class AtsChangeSet extends AbstractAtsChangeSet {
 
    @Override
    public void deleteArtifact(ArtifactId artifact) {
-      getTransaction().deleteArtifact(artifact);
-      add(artifact);
+      if (!atsApi.getStoreService().isDeleted(artifact)) {
+         getTransaction().deleteArtifact(artifact);
+         add(artifact);
+      }
    }
 
    @Override
